refactor: resolve various issues caused by rebase
